chore: use the fancy new create field UI
