+++ frontend/appflowy_flutter/lib/mobile/presentation/database/card/card_detail/widgets/widgets.dart
@@ -1,3 +1,3 @@
 export 'mobile_field_name_text_field.dart';
-export 'mobile_create_row_field_button.dart';
+export 'mobile_create_field_button.dart';
 export 'mobile_row_property_list.dart';
