+++ frontend/appflowy_flutter/lib/mobile/presentation/database/field/bottom_sheet_create_field.dart
@@ -1,0 +1,41 @@
+import 'package:appflowy/mobile/presentation/bottom_sheet/bottom_sheet.dart';
+import 'package:appflowy/mobile/presentation/database/card/card_detail/mobile_create_field_screen.dart';
+import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options.dart';
+import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options_eidtor.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+void showCreateFieldBottomSheet(BuildContext context, String viewId) {
+  showMobileBottomSheet(
+    context,
+    padding: EdgeInsets.zero,
+    builder: (context) {
+      return DraggableScrollableSheet(
+        expand: false,
+        snap: true,
+        initialChildSize: 0.7,
+        minChildSize: 0.7,
+        builder: (context, controller) => FieldOptions(
+          scrollController: controller,
+          onAddField: (type) async {
+            final optionValues = await context.push<FieldOptionValues>(
+              Uri(
+                path: MobileNewPropertyScreen.routeName,
+                queryParameters: {
+                  MobileNewPropertyScreen.argViewId: viewId,
+                  MobileNewPropertyScreen.argFieldTypeId: type.value.toString(),
+                },
+              ).toString(),
+            );
+            if (optionValues != null) {
+              await optionValues.create(viewId: viewId);
+              if (context.mounted) {
+                context.pop();
+              }
+            }
+          },
+        ),
+      );
+    },
+  );
+}
