+++ frontend/appflowy_flutter/lib/plugins/database_view/grid/presentation/widgets/header/grid_header.dart
@@ -1,9 +1,6 @@
 import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
-import 'package:appflowy/mobile/presentation/bottom_sheet/bottom_sheet.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/mobile_create_field_screen.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options_eidtor.dart';
+import 'package:appflowy/mobile/presentation/database/field/bottom_sheet_create_field.dart';
 import 'package:appflowy/plugins/database_view/application/field/field_controller.dart';
 import 'package:appflowy/plugins/database_view/grid/application/grid_bloc.dart';
 import 'package:appflowy/plugins/database_view/grid/application/grid_header_bloc.dart';
@@ -15,7 +12,6 @@ import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
 import 'package:reorderables/reorderables.dart';
 
 import '../../../../application/field/type_option/type_option_service.dart';
@@ -236,7 +232,7 @@ class _CreateFieldButtonState extends State<CreateFieldButton> {
       hoverColor: AFThemeExtension.of(context).greyHover,
       onTap: () async {
         if (PlatformExtension.isMobile) {
-          _showCreateFieldBottomSheet(context);
+          showCreateFieldBottomSheet(context, widget.viewId);
         } else {
           final result = await TypeOptionBackendService.createFieldTypeOption(
             viewId: widget.viewId,
@@ -251,42 +247,6 @@ class _CreateFieldButtonState extends State<CreateFieldButton> {
         FlowySvgs.add_s,
         color: PlatformExtension.isDesktop ? null : Theme.of(context).hintColor,
       ),
-    );
-  }
-
-  void _showCreateFieldBottomSheet(BuildContext context) {
-    showMobileBottomSheet(
-      context,
-      padding: EdgeInsets.zero,
-      builder: (context) {
-        return DraggableScrollableSheet(
-          expand: false,
-          snap: true,
-          initialChildSize: 0.7,
-          minChildSize: 0.7,
-          builder: (context, controller) => FieldOptions(
-            scrollController: controller,
-            onAddField: (type) async {
-              final optionValues = await context.push<FieldOptionValues>(
-                Uri(
-                  path: MobileNewPropertyScreen.routeName,
-                  queryParameters: {
-                    MobileNewPropertyScreen.argViewId: widget.viewId,
-                    MobileNewPropertyScreen.argFieldTypeId:
-                        type.value.toString(),
-                  },
-                ).toString(),
-              );
-              if (optionValues != null) {
-                await optionValues.create(viewId: widget.viewId);
-                if (context.mounted) {
-                  context.pop();
-                }
-              }
-            },
-          ),
-        );
-      },
     );
   }
 }
